+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+* Fix: Change default token type to `asset`
+* Fix: Fix values of `asset` token, fix adding tokens to assets list
+
 ## 0.1.0
 
 Stable development release.